feat(junit): Move table tests into junit-utils module
